Nearly complete rewrite for version 1.0.0
* An info page was added

* The user interface has changed completely:

Now there is a major mode "syntree-mode": the tree is constructed
interactively as the user edits the input string.

The user can define "styles" that specify how the tree is designed.

Now the tree can grow in all four directions.

The dependency on transient.el was dropped because the new user
interface does not use it anymore.