Add new property :compress
If :compress is set to t (as is the case for the predefined styles),
the size of the tree will be optimized (i.e., less wide/tall depending
on whether it grows vertically or horizontally).  This is achieved by
letting nodes be "stacked" on top of each other (in the vertical
growth case) as much as possible.